+++ cpp01/ex06/srcs/Harl.cpp
@@ -67,7 +67,7 @@ void	Harl::error(void) {
 	std::cout << "This is unacceptable! I want to speak to the manager now." << std::endl << std::endl;
 }
 
-ssize_t	Harl::LevelToIndex(std::string& level) {
+int	Harl::LevelToIndex(std::string& level) {
 	for (int i = 0; i < 4; i++) {
 		if ((this->level_list_[i]) == level)
 			return i;
@@ -76,7 +76,7 @@ ssize_t	Harl::LevelToIndex(std::string& level) {
 }
 
 void	Harl::complain(std::string level) {
-	ssize_t		index = LevelToIndex(level);
+	int	index = LevelToIndex(level);
 
 	switch (index) {
 		case 0:
